History
Adding feature to history menu and changing some number in the database

+++ CHKS/Models/Mydb/Connector.cs
@@ -19,7 +19,7 @@ namespace CHKS.Models.mydb
         public Inventory Inventory { get; set; }
 
         [Required]
-        public int Qty { get; set; }
+        public decimal Qty { get; set; }
 
         public string Note { get; set; }
